Add #include required for building with VS2019
Microsoft's VS2019 compiler does not include `runtime_error()` in `stdlib.h`, and thus it explicitly needs to be added with `#include <stdexcept>`.

This has been tested on macOS with XCode and on Windows with VS2019. `stexcept` seems to be a normal part of the C++ lib, http://www.cplusplus.com/reference/stdexcept/, so it is expected that this fix will not break other platforms.